+++ app/src/main/java/com/serrano/academically/room/CourseSkillDao.kt
@@ -21,7 +21,7 @@ interface CourseSkillDao {
     @Query("SELECT courseId FROM CourseSkill WHERE userId = :userId AND role = :role LIMIT 3")
     fun getThreeCourseSkillsOfUserNoRating(userId: Int, role: String): Flow<List<Int>>
 
-    // Usage MessageTutor, FindTutor
+    // Usage MessageTutor, FindTutor, Signup, AssessmentResult
     @Query("SELECT courseId FROM CourseSkill WHERE userId = :userId AND role = :role")
     fun getCourseSkillsOfUserNoRating(userId: Int, role: String): Flow<List<Int>>
 
@@ -29,9 +29,11 @@ interface CourseSkillDao {
     @Query("SELECT * FROM CourseSkill WHERE courseId = :courseId AND role = 'TUTOR'")
     fun getCourseSkillsForTutorByCourse(courseId: Int): Flow<List<CourseSkill>>
 
+    // Usage AssessmentResult
     @Query("SELECT * FROM CourseSkill WHERE courseId = :courseId AND userId = :userId")
     fun getCourseSkill(courseId: Int, userId: Int): Flow<List<CourseSkill>>
 
+    // Usage AssessmentResult
     @Query("UPDATE CourseSkill SET role = :role, courseAssessmentTaken = :taken, courseAssessmentScore = :score, courseAssessmentItemsTotal = :items, courseAssessmentEvaluator = :eval WHERE courseSkillId = :courseSkillId")
-    suspend fun updateCourseSkill(role: String, taken: Int, score: Int, items: Int, eval: Float, courseSkillId: Int)
+    suspend fun updateCourseSkill(role: String, taken: Int, score: Int, items: Int, eval: Double, courseSkillId: Int)
 }
